+++ template-unitybepinex/index.js
@@ -12,8 +12,8 @@ Notes:
 const { actions, fs, util, selectors, log } = require('vortex-api');
 const path = require('path');
 const template = require('string-template');
-const winapi = require('winapi-bindings');
 const { parseStringPromise } = require('xml2js');
+//const winapi = require('winapi-bindings');
 //const fsPromises = require('fs/promises'); //.rm() for recursive folder deletion
 //const fsExtra = require('fs-extra');
 //const turbowalk = require('turbowalk');